renamed model files, updated models, created a table association file called index.js
