+++ model/SidesCategory.js
@@ -1,23 +1,25 @@
 const { Model, DataTypes } = require('sequelize');
 const sequelize = require('../config/config');
 
-class Sauce extends Model {};
+class SidesCategory extends Model {};
 
-Sauce.init({
+SidesCategory.init({
     id: {
         type: DataTypes.INTEGER,
         allowNull: false,
         primaryKey: true,
         autoIncrement: true,
     },
-    name: {
+    title: {
         type: DataTypes.STRING,
         allowNull: false,
-    },
+    }
 },
 {
     sequelize,
-    timestamps: false,
     freezeTableName: true,
-    modelName: 'sauce',
+    underscored: true,
+    modelName: 'category',
 })
+
+module.exports = SidesCategory;
