+++ model/index.js
@@ -1,0 +1,11 @@
+
+const Sides = require('./Side');
+const SidesCategory = require('./SidesCategory');
+
+Sides.belongsTo(SidesCategory, {
+    foreignKey: 'cat_id',
+});
+
+SidesCategory.hasMany(Sides, {
+    foreignKey: 'cat_id',
+});
